add typedoc and scripts to generate API docs under docs/api
