+++ README.md
@@ -4,8 +4,6 @@
 
 Cedar is a library for crafting and sharing data visualizations powered by ArcGIS Services.
 
-> You are looking at the documentation for cedar v1.x, which is currently in beta. You can also view the [v0.x documentation](https://github.com/Esri/cedar/blob/v0.x/README.md).
-
 [View live examples of how to use cedar](http://cedar-v1.surge.sh/)
 
 <!--
