+++ package.json
@@ -8,14 +8,17 @@
     "precopy:docs": "npm run build",
     "clean:deps": "lerna run clean:deps && rimraf node_modules",
     "clean:dist": "lerna run clean:dist",
+    "clean:docs": "yarn run rimraf docs/api docs/scripts",
     "copy:docs": "mkdir -p docs/scripts && cp -r packages/cedar/dist/umd/ docs/scripts/",
     "copy:readme": "cp README.md packages/cedar/README.md",
-    "predeploy": "npm run copy:docs",
+    "predocs": "npm run clean:docs && npm run copy:docs",
+    "docs": "lerna run --scope @esri/cedar docs",
+    "predeploy": "npm run docs",
     "deploy": "surge ./docs cedar-v1.surge.sh",
     "prerelease": "npm run copy:readme && git add .",
     "release": "lerna publish",
     "reset": "npm run clean:dist && npm run clean:deps && yarn",
-    "start": "npm run copy:docs && npm run build:watch",
+    "start": "npm run docs && npm run build:watch",
     "test": "lerna run test"
   },
   "devDependencies": {
