+++ packages/cedar/package.json
@@ -26,7 +26,8 @@
     "test": "jest --coverage",
     "test:watch": "npm run test -- --watch",
     "prepare": "npm run build:prod",
-    "prepublishOnly": "npm test"
+    "prepublishOnly": "npm test",
+    "docs": "yarn run typedoc src/"
   },
   "repository": {
     "type": "git",
@@ -82,6 +83,7 @@
     "snazzy": "^7.0.0",
     "ts-jest": "^23.1.3",
     "tslint": "^5.5.0",
+    "typedoc": "^0.13.0",
     "typescript": "^3.0.1"
   }
 }
