+++ packages/cedar/src/Chart.ts
@@ -1,5 +1,6 @@
 import { cedarAmCharts } from '@esri/cedar-amcharts'
-import { getChartData, IDataset, ISeries } from './dataset'
+import { IDataset, ISeries } from './common'
+import { getChartData } from './dataset'
 import { queryDatasets } from './query/query'
 
 function clone(json) {
@@ -7,16 +8,32 @@ function clone(json) {
 }
 
 // TODO: where should these interfaces live?
+
+/**
+ * Properties to control the visibility and position of the chart's legend
+ */
 export interface ILegend {
   visible?: boolean
   position?: 'top' | 'bottom' | 'left' | 'right'
 }
 
+/**
+ * Properties specific to pie charts
+ */
 export interface IPie {
+  /**
+   * How much of the center of the chart should be cut out to create a donut chart?
+   */
   innerRadius?: number | string
+  /**
+   * Should slices expand outward when clicked?
+   */
   expand?: number | string
 }
 
+/**
+ * Padding around the chart (in pixels)
+ */
 export interface IPadding {
   top?: number
   bottom?: number
@@ -24,12 +41,21 @@ export interface IPadding {
   right?: number
 }
 
+/**
+ * Additional properties to define how a chart is rendered
+ */
 export interface IStyle {
   pie?: IPie
   padding?: IPadding,
+  /**
+   * An array of colors that will be applied to the chart series
+   */
   colors?: string[]
 }
 
+/**
+ * Defines where a chart gets it's data from and how the data is rendered
+ */
 export interface IDefinition {
   datasets?: IDataset[]
   series?: ISeries[]
@@ -40,12 +66,20 @@ export interface IDefinition {
   style?: IStyle
 }
 
+/**
+ * An instance of a cedar chart that will be rendered at a given DOM node (container).
+ */
 export class Chart {
   private _container: string
   private _definition: IDefinition
   private _data: any[]
 
-  constructor(container, definition?) {
+  /**
+   *
+   * @param container The DOM node where the chart will be rendered
+   * @param definition Defines how the data will be rendered in the chart
+   */
+  constructor(container, definition?: IDefinition) {
     if (!container) {
       throw new Error('A container is required')
     }
@@ -58,7 +92,16 @@ export class Chart {
   }
 
   // Setters and getters
+
+  /**
+   *
+   * Set the definition
+   * @param newDefinition
+   */
   public definition(newDefinition: IDefinition): Chart
+  /**
+   * Get the definition
+   */
   public definition(): IDefinition
   public definition(newDefinition?: any): any {
     if (newDefinition === undefined) {
@@ -69,54 +112,109 @@ export class Chart {
     }
   }
 
+  /**
+   *
+   * Set the chart's datasets
+   * @param newDatasets Array of datasets
+   */
   public datasets(newDatasets: IDataset[]): Chart
+  /**
+   * Get the chart's datasets
+   */
   public datasets(): IDataset[]
   public datasets(newDatasets?: any): any {
     return this._definitionAccessor('datasets', newDatasets)
   }
 
+  /**
+   * Set the chart's series
+   * @param newSeries
+   */
   public series(newSeries: ISeries[]): Chart
+  /**
+   * Get the chart's series
+   */
   public series(): ISeries[]
   public series(newSeries?: any): any {
     return this._definitionAccessor('series', newSeries)
   }
 
+  /**
+   * Set the chart type
+   * @param newType
+   */
   public type(newType: string): Chart
+  /**
+   * Get the chart type
+   */
   public type(): string
   public type(newType?: any): any {
     return this._definitionAccessor('type', newType)
   }
 
+  /**
+   * Set the chart's specification
+   * @param newSpecification
+   */
   public specification(newSpecification: {}): Chart
+  /**
+   * Get the chart's specification
+   */
   public specification(): {}
   public specification(newSpecification?: any): any {
     return this._definitionAccessor('specification', newSpecification)
   }
 
+  /**
+   * Override specification properties
+   * @param newOverrides
+   */
   public overrides(newOverrides: {}): Chart
+  /**
+   * Get the specification overrides
+   */
   public overrides(): {}
   public overrides(newOverrides?: any): any {
     return this._definitionAccessor('overrides', newOverrides)
   }
 
+  /**
+   * Set the chart's legend properties
+   * @param newLegend
+   */
   public legend(newLegend: ILegend): Chart
+  /**
+   * Get the chart legend settings
+   */
   public legend(): ILegend
   public legend(newLegend?: any): any {
     return this._definitionAccessor('legend', newLegend)
   }
 
+  /**
+   * Set the chart's styles
+   * @param newStyle
+   */
   public style(newStyle: IStyle): Chart
   public style(): IStyle
+  /**
+   * Get the chart's styles
+   */
   public style(newStyle?: any): any {
     return this._definitionAccessor('style', newStyle)
   }
 
-  // data is read only
+  /**
+   * Get the internal copy of the data used to render the chart
+   */
   public data() {
     return this._data
   }
 
-  // get dataset by name
+  /**
+   * Get a dataset from the definition by name
+   * @param datasetName The name of the dataset to get
+   */
   public dataset(datasetName: string): IDataset {
     const datasets = this.datasets()
     let match
@@ -131,7 +229,9 @@ export class Chart {
     return match
   }
 
-  // query non-inline datasets
+  /**
+   * Query data for all non-inline datasets
+   */
   public query() {
     return queryDatasets(this.datasets())
   }
@@ -147,14 +247,17 @@ export class Chart {
     return this
   }
 
-  // re-draw the chart based on the current state
+  /**
+   * Re-draw the chart based on the current data and definition
+   */
   public render() {
     cedarAmCharts(this._container, this.definition(), this.data())
     return this
   }
 
-  // rollup the query, update, and render functions
-  // useful for showing the chart for the first time
+  /**
+   * Execute the query(), updateData(), and render() functions
+   */
   public show() {
     return this.query()
     .then((response) => {
