+++ packages/cedar/src/index.ts
@@ -1,8 +1,7 @@
 /* Copyright (c) 2018 Environmental Systems Research Institute, Inc.
  * Apache-2.0 */
 
- export * from './Chart'
- export * from './dataset'
- export * from './config'
- export * from './query/query'
- export * from './query/url'
+export * from './Chart'
+export * from './config'
+// types
+export * from './common'
