+++ packages/cedar/typedoc/theme/partials/comment.hbs
@@ -1,0 +1,22 @@
+{{#with comment}}
+    {{#if hasVisibleComponent}}
+        <div class="tsd-comment tsd-typography">
+            {{#if shortText}}
+                <div class="lead">
+                    {{#markdown}}{{{shortText}}}{{/markdown}}
+                </div>
+            {{/if}}
+            {{#if text}}
+                {{#markdown}}{{{text}}}{{/markdown}}
+            {{/if}}
+            {{#if tags}}
+                <dl class="tsd-comment-tags">
+                    {{#each tags}}
+                        <dt>{{tagName}}</dt>
+                        <dd>{{#markdown}}{{{text}}}{{/markdown}}</dd>
+                    {{/each}}
+                </dl>
+            {{/if}}
+        </div>
+    {{/if}}
+{{/with}}
