+++ packages/cedar/typedoc/theme/partials/header.hbs
@@ -1,0 +1,42 @@
+<header>
+    <div class="tsd-page-toolbar">
+        <div class="container">
+            <div class="table-wrap">
+                <div class="table-cell">
+                  <ul>
+                    <li><h1><a href="{{relativeURL "../index.html"}}">{{project.name}}</a></h1></li>
+                    <li class="active"><a href="{{relativeURL "index.html"}}">API</a></li>
+                    <li><a href="https://github.com/Esri/cedar/">GitHub</a></li>
+                  </ul>
+                </div>
+                <div class="table-cell" id="tsd-widgets">
+                    <div id="tsd-filter">
+                        <a href="#" class="tsd-widget options no-caption" data-toggle="options">Options</a>
+                        <div class="tsd-filter-group">
+                            <div class="tsd-select" id="tsd-filter-visibility">
+                                <span class="tsd-select-label">All</span>
+                                <ul class="tsd-select-list">
+                                    <li data-value="public">Public</li>
+                                    <li data-value="protected">Public/Protected</li>
+                                    <li data-value="private" class="selected">All</li>
+                                </ul>
+                            </div>
+
+                            <input type="checkbox" id="tsd-filter-inherited" checked />
+                            <label class="tsd-widget" for="tsd-filter-inherited">Inherited</label>
+
+                            {{#unless settings.excludeExternals}}
+                                <input type="checkbox" id="tsd-filter-externals" checked />
+                                <label class="tsd-widget" for="tsd-filter-externals">Externals</label>
+                            {{/unless}}
+
+                            <input type="checkbox" id="tsd-filter-only-exported" />
+                            <label class="tsd-widget" for="tsd-filter-only-exported">Only exported</label>
+                        </div>
+                    </div>
+                    <a href="#typedoc-main-index" class="tsd-widget menu no-caption">Menu</a>
+                </div>
+            </div>
+        </div>
+    </div>
+</header>
