+++ packages/cedar/typedoc/theme/partials/hierarchy.hbs
@@ -1,0 +1,17 @@
+<ul class="tsd-hierarchy">
+    {{#each types}}
+        <li>
+            {{#if ../isTarget}}
+                <span class="target">{{this}}</span>
+            {{else}}
+                {{> type}}
+            {{/if}}
+
+            {{#if @last}}
+                {{#with ../next}}
+                    {{> hierarchy}}
+                {{/with}}
+            {{/if}}
+        </li>
+    {{/each}}
+</ul>
