+++ packages/cedar/typedoc/theme/partials/index.hbs
@@ -1,0 +1,19 @@
+{{#if groups}}
+<section class="tsd-panel-group tsd-index-group">
+    <h2>Index</h2>
+    <section class="tsd-panel tsd-index-panel">
+        <div class="tsd-index-content">
+            {{#each groups}}
+                <section class="tsd-index-section {{cssClasses}}">
+                    <h3>{{title}}</h3>
+                    <ul class="tsd-index-list">
+                        {{#each children}}
+                        <li class="{{cssClasses}}"><a href="{{relativeURL url}}" class="tsd-kind-icon">{{#if name}}{{{wbr name}}}{{else}}<em>{{{wbr kindString}}}</em>{{/if}}</a></li>
+                        {{/each}}
+                    </ul>
+                </section>
+            {{/each}}
+        </div>
+    </section>
+</section>
+{{/if}}
