+++ packages/cedar/typedoc/theme/partials/member.declaration.hbs
@@ -1,0 +1,22 @@
+<div class="tsd-signature tsd-kind-icon">{{#compact}}
+    {{{wbr name}}}<span class="tsd-signature-symbol">{{#if isOptional}}?{{/if}}:</span>&nbsp;{{#with type}}{{>type}}{{/with}}
+    {{#if defaultValue}}
+        <span class="tsd-signature-symbol">
+        &nbsp;=&nbsp;
+        {{defaultValue}}
+        </span>
+    {{/if}}
+{{/compact}}</div>
+
+{{> member.sources}}
+
+{{> comment}}
+
+{{#if type.declaration}}
+    <div class="tsd-type-declaration">
+        <h4>Type declaration</h4>
+        {{#with type.declaration}}
+            {{> parameter}}
+        {{/with}}
+    </div>
+{{/if}}
