+++ packages/cedar/typedoc/theme/partials/member.hbs
@@ -1,0 +1,27 @@
+<section class="tsd-panel tsd-member {{cssClasses}}">
+    <a name="{{anchor}}" class="tsd-anchor"></a>
+    {{#if name}}
+        <h3>{{#each flags}}<span class="tsd-flag ts-flag{{this}}">{{this}}</span> {{/each}}{{{wbr name}}}</h3>
+    {{/if}}
+
+    {{#if signatures}}
+        {{> member.signatures}}
+    {{else}}
+        {{> member.declaration}}
+    {{/if}}
+
+    {{#unless isContainer}}
+        {{#each groups}}
+            {{#each children}}
+                {{#unless hasOwnDocument}}
+                    {{> member}}
+                {{/unless}}
+            {{/each}}
+        {{/each}}
+    {{/unless}}
+</section>
+
+{{#if isContainer}}
+    {{> index}}
+    {{> members}}
+{{/if}}
