+++ packages/cedar/typedoc/theme/partials/member.signature.title.hbs
@@ -1,0 +1,28 @@
+{{#compact}}
+    {{#unless hideName}}{{{wbr name}}}{{/unless}}
+    {{#if typeParameters}}
+        &lt;
+        {{#each typeParameters}}
+            {{#if @index}},&nbsp;{{/if}}
+            {{name}}
+        {{/each}}
+        &gt;
+    {{/if}}
+    <span class="tsd-signature-symbol">(</span>
+    {{#each parameters}}
+        {{#if @index}},&nbsp;{{/if}}
+        {{#if flags.isRest}}<span class="tsd-signature-symbol">...</span>{{/if}}
+        {{name}}
+        <span class="tsd-signature-symbol">
+            {{#if flags.isOptional}}?{{/if}}
+            {{#if defaultValue}}?{{/if}}
+            :&nbsp;
+        </span>
+        {{#with type}}{{>type}}{{/with}}
+    {{/each}}
+    <span class="tsd-signature-symbol">)</span>
+    {{#if type}}
+        <span class="tsd-signature-symbol">:&nbsp;</span>
+        {{#with type}}{{>type}}{{/with}}
+    {{/if}}
+{{/compact}}
