+++ packages/cedar/typedoc/theme/partials/member.signatures.hbs
@@ -1,0 +1,13 @@
+<ul class="tsd-signatures {{cssClasses}}">
+    {{#each signatures}}
+        <li class="tsd-signature tsd-kind-icon">{{> member.signature.title }}</li>
+    {{/each}}
+</ul>
+
+<ul class="tsd-descriptions">
+    {{#each signatures}}
+        <li class="tsd-description">
+            {{> member.signature.body }}
+        </li>
+    {{/each}}
+</ul>
