+++ packages/cedar/typedoc/theme/partials/member.sources.hbs
@@ -1,0 +1,22 @@
+<aside class="tsd-sources">
+    {{#if implementationOf}}
+        <p>Implementation of {{#with implementationOf}}{{> typeAndParent}}{{/with}}</p>
+    {{/if}}
+    {{#if inheritedFrom}}
+        <p>Inherited from {{#with inheritedFrom}}{{> typeAndParent}}{{/with}}</p>
+    {{/if}}
+    {{#if overwrites}}
+        <p>Overrides {{#with overwrites}}{{> typeAndParent}}{{/with}}</p>
+    {{/if}}
+    {{#if sources}}
+        <ul>
+            {{#each sources}}
+                {{#if url}}
+                    <li>Defined in <a href="{{url}}">{{fileName}}:{{line}}</a></li>
+                {{else}}
+                    <li>Defined in {{fileName}}:{{line}}</li>
+                {{/if}}
+            {{/each}}
+        </ul>
+    {{/if}}
+</aside>
