+++ packages/cedar/typedoc/theme/partials/members.group.hbs
@@ -1,0 +1,8 @@
+<section class="tsd-panel-group tsd-member-group {{cssClasses}}">
+    <h2>{{title}}</h2>
+    {{#each children}}
+        {{#unless hasOwnDocument}}
+            {{> member}}
+        {{/unless}}
+    {{/each}}
+</section>
