+++ packages/cedar/typedoc/theme/partials/members.hbs
@@ -1,0 +1,5 @@
+{{#each groups}}
+    {{#unless allChildrenHaveOwnDocument}}
+        {{> members.group}}
+    {{/unless}}
+{{/each}}
