+++ packages/cedar/typedoc/theme/partials/navigation.hbs
@@ -1,0 +1,26 @@
+{{#if isVisible}}
+    {{#if isLabel}}
+        <li class="label {{cssClasses}}">
+            <span>{{{wbr title}}}</span>
+        </li>
+    {{else}}
+        {{#if isGlobals}}
+            <li class="globals {{#if isInPath}}current{{/if}} {{cssClasses}}">
+                <a href="{{relativeURL url}}"><em>{{{wbr title}}}</em></a>
+            </li>
+        {{else}}
+            <li class="{{#if isInPath}}current{{/if}} {{cssClasses}}">
+                <a href="{{relativeURL url}}">{{{wbr title}}}</a>
+                {{#if isInPath}}
+                    {{#if children}}
+                        <ul>
+                            {{#each children}}
+                                {{> navigation}}
+                            {{/each}}
+                        </ul>
+                    {{/if}}
+                {{/if}}
+            </li>
+        {{/if}}
+    {{/if}}
+{{/if}}
