+++ packages/cedar/typedoc/theme/partials/toc.root.hbs
@@ -1,0 +1,18 @@
+{{#if isInPath}}
+    </ul>
+    <ul class="current">
+{{/if}}
+<li class="{{#if isInPath}}current{{/if}} {{cssClasses}}">
+    <a href="{{relativeURL url}}" class="tsd-kind-icon">{{{wbr title}}}</a>
+    {{#if children}}
+        <ul>
+            {{#each children}}
+                {{> toc}}
+            {{/each}}
+        </ul>
+    {{/if}}
+</li>
+{{#if isInPath}}
+    </ul>
+    <ul class="after-current">
+{{/if}}
