+++ packages/cedar/typedoc/theme/partials/type.hbs
@@ -1,0 +1,83 @@
+{{#if this}}
+    {{#if reflection}}
+        {{#compact}}
+            <a href="{{relativeURL reflection.url}}" class="tsd-signature-type">
+                {{reflection.name}}
+            </a>
+            {{#if typeArguments}}
+                <span class="tsd-signature-symbol">&lt;</span>
+
+                {{#each typeArguments}}
+                    {{#if @index}}
+                        <span class="tsd-signature-symbol">, </span>
+                    {{/if}}{{> type}}
+                {{/each}}
+
+                <span class="tsd-signature-symbol">&gt;</span>
+            {{/if}}
+        {{/compact}}
+    {{else}}
+        {{#if elementType}}
+            {{#with elementType}}
+                {{#compact}}
+                    {{#if types}}
+                        <span class="tsd-signature-symbol">(</span>
+                    {{/if}}
+                    {{> type}}
+                    {{#if types}}
+                        <span class="tsd-signature-symbol">)</span>
+                    {{/if}}<span class="tsd-signature-symbol">[]</span>
+                {{/compact}}
+            {{/with}}
+        {{else}}
+            {{#if types}}
+                {{#each types}}
+                    {{#if @index}}
+                        <span class="tsd-signature-symbol"> {{#ifCond ../type '==' 'intersection'}}&amp;{{else}}|{{/ifCond}} </span>
+                    {{/if}}{{> type}}
+                {{/each}}
+            {{else}}
+                {{#if elements}}
+                    {{#compact}}
+                        <span class="tsd-signature-symbol">[</span>
+
+                        {{#each elements}}
+                            {{#if @index}}
+                                <span class="tsd-signature-symbol">, </span>
+                            {{/if}}{{> type}}
+                        {{/each}}
+
+                        <span class="tsd-signature-symbol">]</span>
+                    {{/compact}}
+                {{else}}
+                    {{#compact}}
+                        <span class="tsd-signature-type">
+                            {{#if name}}
+                                {{name}}
+                            {{else}}
+                                {{#if value}}
+                                    "{{value}}"
+                                {{else}}
+                                    {{this}}
+                                {{/if}}
+                            {{/if}}
+                        </span>
+                        {{#if typeArguments}}
+                            <span class="tsd-signature-symbol">&lt;</span>
+
+                            {{#each typeArguments}}
+                                {{#if @index}}
+                                    <span class="tsd-signature-symbol">, </span>
+                                {{/if}}{{> type}}
+                            {{/each}}
+
+                            <span class="tsd-signature-symbol">&gt;</span>
+                        {{/if}}
+                    {{/compact}}
+                {{/if}}
+            {{/if}}
+        {{/if}}
+    {{/if}}
+{{else}}
+    <span class="tsd-signature-type">void</span>
+{{/if}}
