+++ packages/cedar/typedoc/theme/templates/index.hbs
@@ -1,0 +1,16 @@
+{{#with model}}
+    {{> comment}}
+{{/with}}
+
+{{#if model.typeHierarchy}}
+    <section class="tsd-hierarchy">
+        <h2>Hierarchy</h2>
+        {{#with model.typeHierarchy}}{{> hierarchy}}{{/with}}
+    </section>
+{{/if}}
+
+{{#with model}}
+    <div style="position:relative;"><a name="typedoc-main-index" class="tsd-anchor"></a></div>
+    {{> index}}
+    {{> members}}
+{{/with}}
